implementation of permission mechanism, insight page, refactor

+++ app/controllers/event.js
@@ -1,6 +1,7 @@
 const Event = require('../models/Event')
 const HANDLER = require('../utils/response-helper')
 const HttpStatus = require('http-status-codes')
+const permission = require('../utils/permission')
 const helper = require('../utils/paginate')
 
 module.exports = {
@@ -23,6 +24,7 @@ module.exports = {
       if (!event) {
         return res.status(HttpStatus.BAD_REQUEST).json({ message: 'No post exists' })
       }
+      // check for permission (TODO AFTER PREVIOUS PR MERGED)
       updates.forEach(update => {
         event[update] = req.body[update]
       })
@@ -114,8 +116,11 @@ module.exports = {
       if (!deleteEvent) {
         return res.status(HttpStatus.NOT_FOUND).json({ message: 'No Event exists' })
       }
-      await Event.findByIdAndRemove(id)
-      res.status(HttpStatus.OK).json({ deleteEvent: deleteEvent, message: 'Deleted the event' })
+      if (permission.check(req, res, deleteEvent.createdBy)) {
+        await Event.findByIdAndRemove(id)
+        return res.status(HttpStatus.OK).json({ deleteEvent: deleteEvent, message: 'Deleted the event' })
+      }
+      return res.status(HttpStatus.BAD_REQUEST).json({ msg: 'Not permitted!' })
     } catch (error) {
       HANDLER.handleError(res, error)
     }
